requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.0.0
numpy>=1.24.0
cadquery>=2.3.0

